pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "idcquery"
version = "0.1.0"
description = "IDC BigQuery Query Description helpers for Python"
requires-python =  ">=3.7"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "google-cloud-bigquery",
    "pyyaml",
]
